more item handling

+++ merchant/src/main/java/com/pk/commerce/merchant/api/item/ItemDiscount.java
@@ -2,5 +2,7 @@ package com.pk.commerce.merchant.api.item;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 
-public record ItemDiscount(@JsonValue Integer percent) {
+import java.math.BigDecimal;
+
+public record ItemDiscount(@JsonValue BigDecimal percent) {
 }
